refactor: improve local storage hook code #103 move out header and sidebar components to layouts folder
