+++ src/hooks/useLocalStorage.ts
@@ -1,24 +1,22 @@
 import { useState } from 'react';
+import { LocalStorageService } from '../api/services/localStorageService';
 
 /**
- * Sync state to local storage so that it persists through a page refresh.
- * @param key Local store key.
- * @param initialValue Initial value, returned on errors.
+ * Syncs the component state to the local storage so that it persists through the page refresh.
+ * @param key Local storage key.
+ * @param initialValue Initial value.
  */
 export function useLocalStorage<T>(key: string, initialValue: T): [T, ((value: T) => void)] {
-  // Pass initial state function to useState so logic is only executed once
-  const [storedValue, setStoredValue] = useState<T>(() => {
-    const item = localStorage.getItem(key);
-    return item !== null ? (JSON.parse(item) as T) : initialValue;
-  });
+  const [storedValue, setStoredValue] = useState<T>(LocalStorageService.get<T>(key) ?? initialValue);
 
   /**
-   * Local storage value setter.
+   * Sets a new state value and syncs it with the local storage.
    * @param value Value.
    */
   const setValue = (value: T): void => {
     setStoredValue(value);
-    localStorage.setItem(key, JSON.stringify(value));
+    LocalStorageService.save(key, value);
   };
+
   return [storedValue, setValue];
 }
