+++ src/layouts/DefaultLayout/DefaultLayout.tsx
@@ -1,16 +1,14 @@
 import React, { Suspense, useCallback, useState, VFC } from 'react';
 import { Box } from '@chakra-ui/react';
 import { Outlet } from 'react-router-dom';
-import { Header } from '../../components/Header/Header';
-import { Sidebar } from '../../components/Sidebar/Sidebar';
+import { Header } from '../Header/Header';
+import { Sidebar } from '../Sidebar/Sidebar';
 import { KEY_IS_SIDEBAR_VISIBLE } from '../../utils/localStorageKeys';
 import { useLocalStorage } from '../../hooks';
-import cls from './DefaultLayout.module.css';
 import { Loading } from '../../components';
+import cls from './DefaultLayout.module.css';
 
-/**
- * Default layout with sidebar and header.
- */
+/** Default layout with sidebar and header. */
 export const DefaultLayout: VFC = () => {
   const [isSidebarVisible, setSidebarVisibility] = useLocalStorage(KEY_IS_SIDEBAR_VISIBLE, true);
